+++ src/dataProvider.ts
@@ -493,7 +493,7 @@ async function parseDocumentContent(source: vscode.Uri | vscode.TextDocument) {
         // then with default extension patterns.
         const associations = Object.entries(
             vscode.workspace.getConfiguration('files', uri).get<Record<string, string>>('associations', {}),
-        ).concat([['*.mac', 'spec-data'], ['*.chi', 'chiplot']]);
+        ).concat([['*.spec', 'spec-data'], ['*.chi', 'chiplot']]);
 
         for (const [key, value] of associations) {
             if (minimatch(uri.path, key, { matchBase: true })) {